Split context-menu and example event handler
