+++ src/assets/js/index.js
@@ -1,2 +1,4 @@
 import "./app.js";
 import "./search.js";
+import "./context-menu.js"
+import "./example.js";
